Fix mailcatcher not starting
So, it turns out that the mailcatcher gem is getting installed, but under ruby 1.9.3 and not 2.2.3, which I recently set as the default in order to resolve other installation issues. Because the default gem path is for 2.2.3, it can't find mailcatcher and therefore doesn't start.

I've fixed this in the localvm config by adding the mailcatcher gem to 2.2.3 also (I also had to add another server dev package in order to get it to compile).